:sparkles: 構造体のバイト数を調整してpackingできるようにした

+++ contracts/SalesRecord.sol
@@ -3,5 +3,5 @@ pragma solidity >=0.8.9 <0.9.0;
 
 struct SalesRecord {
     uint8 id;
-    uint256 amount;
+    uint248 amount;
 }
